Separate implementations from interfaces.
