+++ src/main/java/oripa/ORIPA.java
@@ -41,13 +41,13 @@ import oripa.gui.presenter.creasepattern.MouseActionHolder;
 import oripa.gui.presenter.creasepattern.TypeForChangeContext;
 import oripa.gui.presenter.main.MainFramePresenter;
 import oripa.gui.presenter.main.SwitcherBetweenPasteAndChangeOrigin;
-import oripa.gui.view.main.ArrayCopyDialogFactory;
-import oripa.gui.view.main.CircleCopyDialogFactory;
-import oripa.gui.view.main.MainFrame;
-import oripa.gui.view.main.MainFrameSwingDialogFactory;
 import oripa.gui.view.main.MainViewSetting;
-import oripa.gui.view.main.PropertyDialogFactory;
 import oripa.gui.view.main.ViewUpdateSupport;
+import oripa.gui.view.main.swing.ArrayCopyDialogFactory;
+import oripa.gui.view.main.swing.CircleCopyDialogFactory;
+import oripa.gui.view.main.swing.MainFrame;
+import oripa.gui.view.main.swing.MainFrameSwingDialogFactory;
+import oripa.gui.view.main.swing.PropertyDialogFactory;
 import oripa.gui.viewsetting.main.KeyProcessingImpl;
 import oripa.gui.viewsetting.main.MainFrameSettingImpl;
 import oripa.gui.viewsetting.main.PainterScreenUpdater;
