+++ src/main/java/oripa/gui/view/main/swing/DialogWhileFolding.java
@@ -16,7 +16,7 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.gui.view.main;
+package oripa.gui.view.main.swing;
 
 import java.awt.GridBagLayout;
