+++ src/main/java/oripa/gui/view/main/swing/MainFrameSwingDialogFactory.java
@@ -16,9 +16,13 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.gui.view.main;
+package oripa.gui.view.main.swing;
 
 import oripa.gui.view.FrameView;
+import oripa.gui.view.main.ArrayCopyDialogView;
+import oripa.gui.view.main.CircleCopyDialogView;
+import oripa.gui.view.main.MainFrameDialogFactory;
+import oripa.gui.view.main.PropertyDialogView;
 
 /**
  * @author OUCHI Koji
